🩹 [Patch]: Bump `Admin` dependency to v1.1.6 (#45)
## Description

This pull request includes updates to improve code consistency, enhance
functionality, and update documentation for font-related PowerShell
functions. The most important changes include adjustments to argument
completers, module version updates, and corrected documentation links.

### Functional improvements:

*
[`src/functions/completers.ps1`](diffhunk://#diff-d85ead11500a80fca764d18c8480b383cdd0ae75862f268d0d4b4b997b27674bL1-R10):
Refactored the `Register-ArgumentCompleter` block to improve readability
and functionality by using a variable for the scope and returning
results as `CompletionResult` objects.

### Dependency updates:

* `src/functions/public/Get-Font.ps1`,
`src/functions/public/Install-Font.ps1`, and
`src/functions/public/Uninstall-Font.ps1`: Updated the required version
of the `Admin` module from `1.1.5` to `1.1.6`.
[[1]](diffhunk://#diff-ee20084ba781ca437e01d19ac7f613301852186eaaac3f2a2c56227290631228L1-R1)
[[2]](diffhunk://#diff-d3add77d9955c82ca6cb9e9614242dfd502f6692662994dcb98d6a58b923abbfL1-R1)
[[3]](diffhunk://#diff-2bda15615bb104b972c42d08b28af0a1692cbb5ec229ae212fc5959feab71f49L1-R3)

### Documentation fixes:

*
[`src/functions/public/Get-Font.ps1`](diffhunk://#diff-ee20084ba781ca437e01d19ac7f613301852186eaaac3f2a2c56227290631228L72-R72):
Corrected the `.LINK` URL to reflect the updated path for the `Get-Font`
function.
*
[`src/functions/public/Install-Font.ps1`](diffhunk://#diff-d3add77d9955c82ca6cb9e9614242dfd502f6692662994dcb98d6a58b923abbfL104-R104):
Updated the `.LINK` URL for the `Install-Font` function to the correct
path.
*
[`src/functions/public/Uninstall-Font.ps1`](diffhunk://#diff-2bda15615bb104b972c42d08b28af0a1692cbb5ec229ae212fc5959feab71f49L44-R46):
Fixed the `.LINK` URL for the `Uninstall-Font` function to match the new
structure.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas